# swagger/api.yml
swagger: '2.0'
info:
  title: 旅小库微信小程序 后端API接口文档
  version: 1.0.0
  description: 旅小库微信小程序 后端API接口文档
  termsOfService: /tos
  contact:
    name: 旅小库
    url: https://www.lvxiaoku.com.cn
    email: 1243426834@qq.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
schemes:
  - http
  - https
produces:
  - application/json
  - application/x-www-form-urlencoded
consumes:
  - application/x-www-form-urlencoded
...
paths:
...
  /routes/{keyword}:
    get:
      tags:
        - routes
      summary: 搜索路线
      description: 通过关键字搜索相关路线(匹配title、descp)
      operationId: getRouteInfoBySearch
      parameters:
        - name: keyword
          in: path
          description: 搜索关键字
          required: true
          type: string
      responses:
        200:
          description: 搜索路线成功
          schema:
            type: array
            items:
              $ref: '#/definitions/route-preview'
        201:
          description: 搜索路线失败，数据库查询语句错误
          schema:
            $ref: '#/definitions/err'
  /routes/edit/new:
    post:
      tags:
        - routeEdit
      summary: 新建路线
      description: 新建路线，必须的参数是token，其他参数是新建路线信息,返回新建路线的id
      operationId: createNewRoute
      parameters:
        - name: data
          in: body
          required: true
          schema:
            type: object
            properties:
              token:
                type: string
                description: 用户token，32位
                example: 1234567890123456789012
              title:
                type: string
                description: 路线标题
                example: 北京到上海的旅程
              descp:
                type: string
                description: 路线描述
                example: 北京到上海的旅程，从早上6点出发，途经多个景点，最后到达上海虹桥机场
              ispub:
                type: integer
                description: 是否公开，1为公开，0为私有
                example: 1
              daynum:
                type: integer
                description: 路线天数
                example: 3
              locnum:
                type: integer
                description: 路线总地点数
                example: 5
              tag:
                type: integer
                description: 路线标签
                example: 1
      responses:
        200:
          description: 新建路线成功
          schema:
            type: object
            properties:
              msg:
                type: string
                description: 新建路线成功提示信息
                example: create new route success
              route_id:
                type: integer
                description: 所新建路线的专属id，根据此id进行路线编辑、删除等操作
                example: 1001
        201:
          description: 新建路线失败，数据库查询语句错误
          schema:
            $ref: '#/definitions/err'
        202:
          description: 新建路线失败，缺少token参数
          schema:
            $ref: '#/definitions/err'
          examples:
            application/json:
              errmsg: 'Missing parameters: token'
              errcode: -2
  /routes/edit/update:
    put:
      tags:
        - routeEdit
      summary: 更新路线详细信息
      description: 更新路线详细信息，必须携带路线route_id，其他参数是更新条目
      operationId: updateRouteInfo
      parameters:
        - name: data
          in: body
          required: true
          schema:
            type: object
            properties:
              route_id:
                type: integer
                description: 路线id,必须要有的参数,其他参数可以选择
                example: 1001
              title:
                type: string
                description: 路线标题
                example: 北京到上海的旅程
              descp:
                type: string
                description: 路线描述
                example: 北京到上海的旅程，从早上6点出发，途经多个景点，最后到达上海虹桥机场
              ispub:
                type: integer
                description: 是否公开，1为公开，0为私有
                example: 1
              daynum:
                type: integer
                description: 路线天数
                example: 3
              locnum:
                type: integer
                description: 路线总地点数
                example: 5
              tag:
                type: integer
                description: 路线标签
                example: 1
      responses:
        200:
          description: 更新路线成功
          schema:
            type: object
            properties:
              msg:
                type: string
                description: 更新路线成功提示信息
                example: update route success
        201:
          description: 更新路线失败，数据库查询语句错误
          schema:
            $ref: '#/definitions/err'
        202:
          description: '更新路线失败，缺少参数route_id'
          schema:
            $ref: '#/definitions/err'
          examples:
            application/json:
              errmsg: 'Missing parameters: route_id'
              errcode: -2
    delete:
      summary: 删除此条路线
      description: 根据路线route_id，删除此条路线
      operationId: deleteRoute
      tags:
        - routeEdit
      parameters:
        - name: route_id
          in: formData
          description: 路线id
          required: true
          type: integer
      responses:
        200:
          description: 删除成功
          schema:
            type: object
            properties:
              msg:
                type: string
                description: 更删除路线成功提示信息
                example: delete route success
        201:
          description: 删除失败，数据库查询语句错误
          schema:
            $ref: '#/definitions/err'
        202:
          description: '删除失败，缺少参数route_id'
          schema:
            $ref: '#/definitions/err'
          examples:
            application/json:
              errmsg: 'Missing parameters: route_id'
              errcode: -2
  /routes/edit/location:
    post:
      tags:
        - routeEdit
      summary: 新增路线地点
      description: 新增路线地点，参数为路线id和地点信息
      operationId: addRouteLocation
      consumes:
        - application/json
      parameters:
        - name: data
          in: body
          description: 路线id和data,两个参数必须要有
          required: true
          schema:
            type: object
            properties:
              route_id:
                type: integer
                description: 路线id
                example: 1001
              data:
                $ref: '#/definitions/location-list'
      responses:
        200:
          description: 新增路线地点成功
          schema:
            type: object
            properties:
              msg:
                type: string
                description: 新增路线地点成功提示信息
                example: insert new locations success
        201:
          description: 新增路线地点失败，数据库查询语句错误
          schema:
            $ref: '#/definitions/err'
        202:
          description: '新增路线地点失败，缺少参数route_id或者参数data'
          schema:
            $ref: '#/definitions/err'
          examples:
            application/json:
              errmsg: 'Missing parameters: route_id'
              errcode: -2
    put:
      tags:
        - routeEdit
      summary: 更新路线地点信息
      description: 更新路线地点信息，参数为路线id和需要更新的完整地点信息
      operationId: updateRouteLocation
      consumes:
        - application/json
      parameters:
        - name: data
          in: body
          description: 路线id
          required: true
          schema:
            type: object
            properties:
              route_id:
                type: integer
                description: 路线id
                example: 1001
              data:
                $ref: '#/definitions/location-list'
      responses:
        200:
          description: 更新路线地点信息成功
          schema:
            type: object
            properties:
              msg:
                type: string
                description: 更新路线地点信息成功提示信息
                example: update locations success
        201:
          description: 更新路线地点信息失败，数据库查询语句错误
          schema:
            $ref: '#/definitions/err'
        202:
          description: '更新路线地点信息失败，缺少参数route_id或者参数data'
          schema:
            $ref: '#/definitions/err'
          examples:
            application/json:
              errmsg: 'Missing parameters: route_id'
              errcode: -2
    delete:
      tags:
        - routeEdit
      summary: 删除路线地点
      description: 删除路线地点，参数为路线id和地点id
      operationId: deleteRouteLocation
      consumes:
        - application/json
      parameters:
        - name: data
          in: body
          required: true
          schema:
            type: object
            properties:
              route_id:
                type: integer
                description: 路线id
                example: 1001
              data:
                type: array
                items:
                  type: string
                  description: 地点名字
                  example: ['南京路', '人民广场']
      responses:
        200:
          description: 删除路线地点成功
          schema:
            type: object
            properties:
              msg:
                type: string
                description: 删除路线地点成功提示信息
                example: delete location success
        201:
          description: 删除路线地点失败，数据库查询语句错误
          schema:
            $ref: '#/definitions/err'
        202:
          description: '删除路线地点失败，缺少参数route_id或者参数loc_id'
          schema:
            $ref: '#/definitions/err'
          examples:
            application/json:
              errmsg: 'Missing parameters: route_id'
              errcode: -2
  /routes/edit/way:
    post:
      tags:
        - routeEdit
      summary: 新增路线规划
      description: 新增路线规划，参数为路线id和data(表示路线规划信息)
      operationId: addRouteWay
      consumes:
        - application/json
      parameters:
        - name: data
          in: body
          description: 路线id和data,两个参数必须要有
          required: true
          schema:
            type: object
            properties:
              route_id:
                type: integer
                description: 路线id
                example: 1001
              data:
                description: 路线规划信息
                $ref: '#/definitions/way-list'
      responses:
        200:
          description: 新增路线规划成功
          schema:
            type: object
            properties:
              msg:
                type: string
                description: 新增路线规划成功提示信息
                example: insert new way success
        201:
          description: 新增路线规划失败，数据库查询语句错误
          schema:
            $ref: '#/definitions/err'
        202:
          description: '新增路线规划失败，缺少参数route_id或者参数data'
          schema:
            $ref: '#/definitions/err'
          examples:
            application/json:
              errmsg: 'Missing parameters: route_id'
              errcode: -2
    put:
      tags:
        - routeEdit
      summary: 更新路线规划信息
      description: 更新路线规划信息，参数为路线id和需要更新的完整路线规划信息
      operationId: updateRouteWay
      consumes:
        - application/json
      parameters:
        - name: data
          in: body
          description: 路线id
          required: true
          schema:
            type: object
            properties:
              route_id:
                type: integer
                description: 路线id
                example: 1001
              data:
                description: 路线规划信息
                $ref: '#/definitions/way-list'
      responses:
        200:
          description: 更新路线规划信息成功
          schema:
            type: object
            properties:
              msg:
                type: string
                description: 更新路线规划信息成功提示信息
                example: 'update way success'
        201:
          description: 更新路线规划信息失败，数据库查询语句错误
          schema:
            $ref: '#/definitions/err'
        202:
          description: '更新路线规划信息失败，缺少参数route_id或者参数data'
          schema:
            $ref: '#/definitions/err'
          examples:
            application/json:
              errmsg: 'Missing parameters: route_id'
              errcode: -2
  /routes/edit/traffic:
    post:
      tags:
        - routeEdit
      summary: 新增路线交通信息
      description: 新增路线交通信息，参数为路线id和data(表示交通信息)
      operationId: addRouteTraffic
      consumes:
        - application/json
      parameters:
        - name: data
          in: body
          description: 路线id和data,两个参数必须要有
          required: true
          schema:
            type: object
            properties:
              route_id:
                type: integer
                description: 路线id
                example: 1001
              data:
                description: 路线交通信息
                $ref: '#/definitions/traffic-list'
      responses:
        200:
          description: 新增路线交通信息成功
          schema:
            type: object
            properties:
              msg:
                type: string
                description: 新增路线交通信息成功提示信息
                example: 'insert new traffic success'
        201:
          description: 新增路线交通信息失败，数据库查询语句错误
          schema:
            $ref: '#/definitions/err'
        202:
          description: '新增路线交通信息失败，缺少参数route_id或者参数data'
          schema:
            $ref: '#/definitions/err'
          examples:
            application/json:
              errmsg: 'Missing parameters: route_id'
              errcode: -2
    put:
      tags:
        - routeEdit
      summary: 更新路线交通信息
      description: 更新路线交通信息，参数为路线id和需要更新的完整路线交通信息，更新前的交通信息会被删除
      operationId: updateRouteTraffic
      consumes:
        - application/json
      parameters:
        - name: data
          in: body
          description: 路线id
          required: true
          schema:
            type: object
            properties:
              route_id:
                type: integer
                description: 路线id
                example: 1001
              data:
                description: 路线交通信息
                $ref: '#/definitions/traffic-list'
      responses:
        200:
          description: 更新路线交通信息成功
          schema:
            type: object
            properties:
              msg:
                type: string
                description: 更新路线交通信息成功提示信息
                example: 'update traffic success'
        201:
          description: 更新路线交通信息失败，数据库查询语句错误
          schema:
            $ref: '#/definitions/err'
        202:
          description: '更新路线交通信息失败，缺少参数route_id或者参数data'
          schema:
            $ref: '#/definitions/err'
          examples:
            application/json:
              errmsg: 'Missing parameters: route_id'
              errcode: -2
  /score/route:
    get:
      tags:
        - score
      summary: 获取路线下的所有评分
      description: 获取路线下的所有评分，正确返回是一个数组，每个元素是一条评分详细信息
      operationId: getRoutesScores
      parameters:
        - name: data
          in: body
          required: true
          description: 参数为路线route_id
          schema:
            type: object
            properties:
              route_id:
                type: integer
                description: 路线id
                example: 1001
      responses:
        200:
          description: 获取成功，返回路线下的所有评分
          schema:
            type: array
            items:
              $ref: '#/definitions/score-detail'
        201:
          description: "获取失败，数据库查询语句错误"
          schema:
            $ref: '#/definitions/err'
        202:
          description: "获取失败，缺少参数route_id"
          schema:
            $ref: '#/definitions/err'
          examples:
            application/json:
              errmsg: 'Missing parameters: route_id'
              errcode: -2
...
definitions:
...
  err:
    type: object
    properties:
      errmsg:
        type: string
        description: 错误信息
        example: 'Database query failed'
      errcode:
        type: integer
        description: 错误码
        example: -1
...
  route-preview:
    type: object
    properties:
      route_id:
        type: integer
        description: 路线id
        example: 1001
      title:
        type: string
        description: 路线标题
        example: 上海三日游
      token:
        type: string
        description: 路线拥有者token
        example: c55beb693aeaea300f7c7344b7cf7ecd
      avatar:
        type: string
        description: 路线作者头像下载url
        example: /download/img/avatar/default.jpeg
      cover:
        type: string
        description: 路线封面下载url
        example: /download/img/cover/default.jpeg
      descp:
        type: string
        description: 路线描述
        example: 三日游，去上海玩，遇见你
      daynum:
        type: integer
        description: 路线天数
        example: 3
      locnum:
        type: integer
        description: 路线总地点数
        example: 10
      likenum:
        type: integer
        description: 路线点赞数
        example: 100
      fork:
        type: integer
        description: 路线被fork数
        example: 10
      score:
        type: number
        description: 路线平均评分，一位小数
        example: 4.5
      pingfen:
        type: number
        description: 路线得分
        example: 563.3
...
  location:
    type: object
    properties:
      name:
        type: string
        description: 地点名称
        example: 人民广场
      "address":
        type: string
        description: 地点地址
        example: "上海市黄浦区上海人民广场"
      "latitude":
        type: number
        description: 地点纬度
        example: 31.229381
      "longitude":
        type: number
        description: 地点经度
        example: 121.474539
  location-list:
    type: array
    description: 路线地点列表,元素为每一天的地点列表
    items:
      type: array
      description: 每一天的地点列表，元素为具体每个地点的信息
      items:
        $ref: '#/definitions/location'
  way-list:
    type: array
    description: 路线规划,元素为每一天的路线规划list
    items:
      type: array
      description: 每一天的路线规划list，元素为具体每个路线途径地点信息
      items:
        $ref: '#/definitions/location'
      example:
        - name: 人民广场
          address: 上海市黄浦区上海人民广场
          latitude: 31.2294
          longitude: 121.4755
          round_time: 60
        - name: 老场坊
          address: 上海市浦东新区世纪大道100号
          latitude: 31.2546
          longitude: 121.492
          round_time: 60
        - name: 甜爱路
          address: 上海市虹口区甜爱路与甜爱支路交叉口南60米
          latitude: 31.2676
          longitude: 121.484
          round_time: 60
  traffic:
    type: object
    properties:
      approach:
        type: integer
        description: 交通方式，0为驾车，1为步行，2为地铁，3为公交
        example: 2
      line:
        type: string
        description: 交通线路描述
        example: 地铁8号线
      start:
        type: string
        description: 交通起点
        example: 人民广场站
      ed:
        type: string
        description: 交通终点
        example: 曲阜路站
      tme:
        type: integer
        description: 交通时长
        example: 30
  traffic-list:
    type: array
    description: 路线交通,元素为每一天的路线交通信息
    items:
      type: array
      description: 每一天的路线交通信息，元素为具体每一段路线交通信息
      items:
        type: array
        description: 具体某一段路线交通信息，元素为交通规划
        items:
          $ref: '#/definitions/traffic'
  score-detail:
    type: object
    properties:
      token:
        type: string
        description: 用户token
        example: c55beb693aeaea300f7c7344b7cf7ecd
      avatar:
        type: string
        description: 用户头像下载url
        example: /download/img/avatar/default.jpeg
      nkname:
        type: string
        description: 用户昵称
        example: 张三
      uptime:
        type: string
        description: 评分时间
        example: '2024-03-06 00:00:00'
      lscore:
        type: number
        description: 路线中地点列表评分，一位小数
        example: 4.5
      rscore:
        type: number
        description: 路线规划评分，一位小数
        example: 4.5
      tscore:
        type: number
        description: 路线交通评分，一位小数
        example: 4.5
...